DownSampling the k space.
eg, Use a 8192 k-space box for a 16384 k-space, with twice the
k-separation. saves memory.

The real space box is therefore twice smaller.
The smaller box is replicated to fill the entire space.

The random seeds/seq are the same as the full 16384 sequence, only taking
these at the even positions.

The dumps from ngenic.grid is still on the full 16384 resolution;
no changes in convert.py is required.