catted Outputs files are irrel from num of MPIs, even with DownSample
without DownSampling, there is only one box and the old logic is already
the case. the file name is xxxxx.%0*d. the number is the task number.

with DownSampling, each x-duplicate strip in the real space is dump to a individual
file, so that after the files are catted along the file ids the final result
is always the same. The file name is xxxxx.%0*d-%0*d. The first number is the
x-duplicate strip number and the second is the task number.