attic code for the new OctTree and particle storage code.
need to redo the load balancing stuff.
inserting/removing particles on the fly seems to be alright

packing/unpacking appears fine. so we can pack a node and send it off
to another processor.

need to remove a node. also to clear up a node after it no longer contains
any external non-empty nodes.